requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==25.1
djangorestframework==3.16.1
importlib_metadata==8.7.0
Markdown==3.9
pillow==11.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.15.0
zipp==3.23.0
